+++ CHANGELOG.md
@@ -14,6 +14,12 @@
 
 
 
+### DEPRECATIONS:
+
+* The `in` parameter of `session.Exec` is deprecated and will be removed in a future release. Please pass body within `*http.Request` instead.
+
+
+
 
 
 
@@ -36,6 +42,7 @@
 * Fixed incorrect Authorization header generation for POST requests when `Config.MaxBody` was not set (defaulted to 0). 
   `SignRequest` now defaults `MaxBody` to `MaxBodySize` (131072) when the value is 0 or negative. 
   ([I#234](https://github.com/akamai/AkamaiOPEN-edgegrid-golang/issues/234)).
+